♻️ Move TOC extraction logic to TableOfContents/utils
- Extract parseHeading and extractTocItems functions to dedicated utils
- Move TocItem type definition alongside the functions
- Keep TOC-specific logic within TableOfContents component directory
- Maintain separation between UI component and business logic

This improves component self-containment by keeping related utilities
within the component's own directory structure.
