+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/Artifact/TableOfContents/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './extractTocItems'
